230524 정승현 개발완료

+++ src/main/java/egovframework/diam/web/main/WebSubscribeController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 
 import egovframework.diam.biz.model.main.Dm_subscribe_vo;
 import egovframework.diam.biz.service.main.SubscribeService;
+import egovframework.diam.cmm.util.CommonUtil;
 import egovframework.diam.cmm.util.MessageCode;
 import lombok.extern.log4j.Log4j2;
 
@@ -24,8 +25,12 @@ public class WebSubscribeController {
 	@PostMapping("/web/requestSubscribe.do")
 	public ResponseEntity<?> requestSubscribe(Dm_subscribe_vo vo) {
 		Map<String, Object> resultMap = new HashMap<>();
-		
+		CommonUtil commonUtil = new CommonUtil();
 		try {
+			if (commonUtil.isNullOrEmpty(vo.getDm_name()) || commonUtil.isNullOrEmpty(vo.getDm_email()) || commonUtil.isNullOrEmpty(vo.getDm_status())) {
+				resultMap.put("notice", MessageCode.CMM_REQUEST_BADREQUEST.getMessage());
+				return new ResponseEntity<>(resultMap, HttpStatus.BAD_REQUEST);
+			}
 			
 			Dm_subscribe_vo chk = subscribeService.selectByEmail(vo);
 			
@@ -49,14 +54,19 @@ public class WebSubscribeController {
 					resultMap.put("result", "fail");
 					resultMap.put("notice", "입력 정보는 이미 구독중입니다.");
 				} else {
-					vo.setDm_id(chk.getDm_id());
-					int result = subscribeService.deleteSubscribe(vo);
-					if (result > 0) {
-						resultMap.put("result", "success");
-						resultMap.put("notice", "구독 해지 되었습니다.");
-					} else {
+					if (!chk.getDm_name().equals(vo.getDm_name())) {
 						resultMap.put("result", "fail");
-						resultMap.put("notice", MessageCode.CMS_UPDATE_FAIL.getMessage());
+						resultMap.put("notice", "입력한 이메일과 이름이 일치하지 않습니다.");
+					} else {
+						vo.setDm_id(chk.getDm_id());
+						int result = subscribeService.deleteSubscribe(vo);
+						if (result > 0) {
+							resultMap.put("result", "success");
+							resultMap.put("notice", "구독 해지 되었습니다.");
+						} else {
+							resultMap.put("result", "fail");
+							resultMap.put("notice", MessageCode.CMS_UPDATE_FAIL.getMessage());
+						}
 					}
 				}
 			}
